# Make fmt::Display for Isbn10 / Isbn13 4x faster