Instead of writing each digit independently, we can collect them into an ArrayString and write all at once, reducing the number of write! calls significantly, and preventing all intermediate errors.

The times are as follows:

Before:
```
test tests::test_isbn10_to_str ... bench:         358 ns/iter (+/- 19)
test tests::test_isbn13_to_str ... bench:         409 ns/iter (+/- 27)
```

After:
```
test tests::test_isbn10_to_str ... bench:         100 ns/iter (+/- 1)
test tests::test_isbn13_to_str ... bench:         109 ns/iter (+/- 9)
```